+++ src/coap.h
@@ -384,7 +384,7 @@ coap_error coap_set_token(coap_pdu *pdu, uint64_t token, uint8_t tkl);
 /// @param  [in]      opt  option container.
 /// @return coap_error (0 == no error)
 ///
-coap_error coap_add_option(coap_pdu *pdu, int32_t opt_num, uint8_t* value, uint16_t opt_len);
+coap_error coap_add_option(coap_pdu *pdu, uint16_t opt_num, uint8_t* value, uint16_t opt_len);
 
 ///
 /// Add Message Option
